Ripple effect of refactoring duplicate code in SVGEllipseFigure
